+++ Gateway/Http/Client/Rest.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright (c) 2016-2019 Mastercard
+ * Copyright (c) 2016-2022 Mastercard
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,6 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 namespace OnTap\MasterCard\Gateway\Http\Client;
 
 use Magento\Framework\Serialize\Serializer\Json;
@@ -22,6 +23,7 @@ use Magento\Payment\Gateway\Http\ClientInterface;
 use Magento\Payment\Gateway\Http\ConverterInterface;
 use Magento\Payment\Gateway\Http\TransferInterface;
 use Magento\Payment\Model\Method\Logger;
+use Zend_Http_Client_Adapter_Interface;
 
 class Rest implements ClientInterface
 {
@@ -34,16 +36,16 @@ class Rest implements ClientInterface
     /**
      * HTTP request methods
      */
-    const GET     = 'GET';
-    const POST    = 'POST';
-    const PUT     = 'PUT';
-    const HEAD    = 'HEAD';
-    const DELETE  = 'DELETE';
-    const TRACE   = 'TRACE';
+    const GET = 'GET';
+    const POST = 'POST';
+    const PUT = 'PUT';
+    const HEAD = 'HEAD';
+    const DELETE = 'DELETE';
+    const TRACE = 'TRACE';
     const OPTIONS = 'OPTIONS';
     const CONNECT = 'CONNECT';
-    const MERGE   = 'MERGE';
-    const PATCH   = 'PATCH';
+    const MERGE = 'MERGE';
+    const PATCH = 'PATCH';
 
     /**
      * @const int Request timeout
@@ -66,28 +68,27 @@ class Rest implements ClientInterface
     private $responseFactory;
 
     /**
-     * @var \Zend_Http_Client_Adapter_Interface
+     * @var Zend_Http_Client_Adapter_Interface
      */
     private $adapter;
+
     /**
      * @var Json
      */
     private $json;
 
     /**
-     * Constructor
-     *
      * @param Logger $logger
      * @param ConverterInterface $converter
      * @param ResponseFactory $responseFactory
-     * @param \Zend_Http_Client_Adapter_Interface $adapter
+     * @param Zend_Http_Client_Adapter_Interface $adapter
      * @param Json $json
      */
     public function __construct(
         Logger $logger,
         ConverterInterface $converter,
         ResponseFactory $responseFactory,
-        \Zend_Http_Client_Adapter_Interface $adapter,
+        Zend_Http_Client_Adapter_Interface $adapter,
         Json $json
     ) {
         $this->logger = $logger;
@@ -102,23 +103,23 @@ class Rest implements ClientInterface
      */
     public function placeRequest(TransferInterface $transferObject)
     {
-        $log = [
-            'request' => json_encode($transferObject->getBody(), JSON_UNESCAPED_SLASHES),
-            'request_uri' => $transferObject->getUri()
-        ];
         $response = [];
 
         try {
             $this->adapter->setOptions(
-                [
-                    CURLOPT_USERPWD => $transferObject->getAuthUsername() . ":" . $transferObject->getAuthPassword(),
-                    CURLOPT_TIMEOUT => self::REQUEST_TIMEOUT
-                ]
+                [CURLOPT_TIMEOUT => self::REQUEST_TIMEOUT] + $transferObject->getClientConfig()
             );
             $headers = [];
+
             foreach ($transferObject->getHeaders() as $name => $value) {
                 $headers[] = sprintf('%s: %s', $name, $value);
             }
+
+            $log = [
+                'request' => json_encode($transferObject->getBody(), JSON_THROW_ON_ERROR | JSON_UNESCAPED_SLASHES),
+                'request_uri' => $transferObject->getUri(),
+            ];
+
             $this->adapter->write(
                 $transferObject->getMethod(),
                 \Zend_Uri_Http::fromString($transferObject->getUri()),
@@ -135,7 +136,7 @@ class Rest implements ClientInterface
             $this->logger->debug($log);
         }
 
-        return (array) $response;
+        return (array)$response;
     }
 
     /**