+++ Gateway/Response/OrderTokenCreateHandler.php
@@ -1,0 +1,67 @@
+<?php
+/**
+ * Copyright (c) 2016-2022 Mastercard
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace OnTap\MasterCard\Gateway\Response;
+
+use Magento\Payment\Gateway\Helper\SubjectReader;
+use Magento\Payment\Gateway\Response\HandlerInterface;
+use Magento\Sales\Api\Data\OrderExtensionFactory;
+use Magento\Sales\Model\Order\Payment;
+
+class OrderTokenCreateHandler implements HandlerInterface
+{
+    /**
+     * @var OrderExtensionFactory
+     */
+    private $orderExtensionFactory;
+
+    /**
+     * @param OrderExtensionFactory $orderExtensionFactory
+     */
+    public function __construct(OrderExtensionFactory $orderExtensionFactory)
+    {
+        $this->orderExtensionFactory = $orderExtensionFactory;
+    }
+
+    /**
+     * @param array $handlingSubject
+     * @param array $response
+     *
+     * @return void
+     */
+    public function handle(array $handlingSubject, array $response)
+    {
+        $paymentToken = $response['token'] ?? null;
+        if (!$paymentToken) {
+            return;
+        }
+
+        $paymentDO = SubjectReader::readPayment($handlingSubject);
+
+        /** @var Payment $payment */
+        $payment = $paymentDO->getPayment();
+
+        $order = $payment->getOrder();
+        $order->setData('mastercard_payment_token', $paymentToken);
+        $extensionAttributes = $order->getExtensionAttributes();
+        if (!$extensionAttributes) {
+            $extensionAttributes = $this->orderExtensionFactory->create();
+        }
+        $extensionAttributes->setMastercardPaymentToken($paymentToken);
+        $order->setExtensionAttributes($extensionAttributes);
+    }
+}