+++ Gateway/Validator/TokenCreateValidator.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright (c) 2016-2019 Mastercard
+ * Copyright (c) 2016-2022 Mastercard
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -57,7 +57,7 @@ class TokenCreateValidator extends AbstractValidator
             return $this->createResult(false, [$response['error']['explanation']]);
         }
 
-        if ($response['status'] == static::STATUS_VALID) {
+        if ($response['status'] === static::STATUS_VALID) {
             return $this->createResult(true);
         }
         return $this->createResult(false, ['Failed to tokenize card']);